Memoized neighbor PID lookup
No need to constantly call into the registry. We can do it one on the
first update we see it is missing and cache the PIDs. Significant
reduction in time cost when dealing with large (> 100x100) boards. Can't
get much larger than that since we run into process limits and execution
time limits on my little machine.